+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 <!---
-SPDX-FileCopyrightText: 2020 Robin Krahl <robin.krahl@ireas.org>
+SPDX-FileCopyrightText: 2020-2021 Robin Krahl <robin.krahl@ireas.org>
 SPDX-License-Identifier: MIT
 -->
 
@@ -9,6 +9,7 @@ SPDX-License-Identifier: MIT
 
 - Add `o` command for opening a documentation item to the tui viewer.
 - Add tests for Rust 1.48.0, 1.49.0, 1.50.0, 1.51.0, 1.52.0 and 1.52.1.
+- Add support for the new search index format introduced in Rust 1.52.0.
 
 ## v0.4.1 (2020-10-11)